+++ autoload/omnipytent/__init__.py
@@ -1,8 +1,21 @@
-__all__ = ['task', 'CMD', 'FN', 'VAR', 'OPT', 'BANG', 'SH', 'INPUT_BUFFER', 'CHOOSE']
+__all__ = [
+    'task',
+    'CMD',
+    'FN',
+    'VAR',
+    'OPT',
+    'BANG',
+    'SH',
+    'INPUT_BUFFER',
+    'CHOOSE',
+    'CombineSources',
+    'CombineSourcesMulti',
+]
 
 from .dsl import task
 from .execution import CMD, FN, VAR, OPT, BANG, SH
 from .async_execution import INPUT_BUFFER, CHOOSE
+from .task_makers import CombineSources, CombineSourcesMulti
 
 from . import _extending as __extending